Got everything refactored to fit with the project
There are still errors though
