+++ CardGameSuiteApp/app/src/main/AndroidManifest.xml
@@ -8,13 +8,20 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.CardGameSuiteApp">
-        <activity android:name=".gameCollectionMainMenu.DisplaySettingsActivity"></activity>
-        <activity android:name=".games.Fives"></activity>
-        <activity android:name=".gameCollectionMainMenu.DisplayFivesRulesActivity"></activity>
-        <activity android:name=".games.Solitaire"></activity>
-        <activity android:name=".gameCollectionMainMenu.DisplaySolitaireRulesActivity"></activity>
-        <activity android:name=".gameCollectionMainMenu.DisplayMainPageActivity" android:screenOrientation="portrait">
+        android:theme="@style/Theme.CardGameSuiteApp"
+        android:usesCleartextTraffic="true">
+        <activity android:name=".gameCollectionMainMenu.DisplaySettingsActivity" />
+        <activity android:name=".games.Fives" />
+        <activity android:name=".gameCollectionMainMenu.DisplayFivesRulesActivity" />
+        <activity android:name=".games.Solitaire" />
+        <activity android:name=".gameCollectionMainMenu.DisplaySolitaireRulesActivity" />
+        <activity
+            android:name=".austenMPStuff.MultiplayerWaitingRoomActivity"
+            android:label="@string/title_activity_multiplayer_waiting_room"
+            android:parentActivityName=".austenMPStuff.MainActivity" />
+        <activity
+            android:name=".gameCollectionMainMenu.DisplayMainPageActivity"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
