+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/austenMPStuff/IMultiplayerConnectorSocketEventUser.java
@@ -1,4 +1,4 @@
-package com.example.cardgamesuiteapp;
+package com.example.cardgamesuiteapp.austenMPStuff;
 
 import io.socket.client.Socket;
