+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/austenMPStuff/MultiplayerWaitingRoomActivityFragment.java
@@ -1,4 +1,4 @@
-package com.example.cardgamesuiteapp;
+package com.example.cardgamesuiteapp.austenMPStuff;
 
 import android.os.Bundle;
 import android.view.View;
