+++ CardGameSuiteApp/app/src/main/res/layout/austen_activity_multiplayer_waiting_room.xml
@@ -5,7 +5,7 @@
     android:id="@+id/multiPlayerWaitingRoomCoordinatorLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MultiplayerWaitingRoomActivity">
+    tools:context=".austenMPStuff.MultiplayerWaitingRoomActivity">
 
 
     <RelativeLayout
