+++ CardGameSuiteApp/app/src/main/res/layout/austen_content_main.xml
