+++ CardGameSuiteApp/app/src/main/res/layout/austen_fragment_first.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FirstFragment">
+    tools:context=".austenMPStuff.FirstFragment">
 
     <Button
         android:id="@+id/button_first"
