+++ CardGameSuiteApp/app/src/main/res/layout/austen_fragment_second2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Second2Fragment">
+    tools:context=".austenMPStuff.Second2Fragment">
 
     <Button
         android:id="@+id/button_second"
